:refactor: refactor: [KAN-108] 로직 수정으로 인해 로직 축소 및 파일명 변경
